modelling of plastic proportions 01_04_2020
After talking with Stuart I have changed some things regarding the modelling of the proportions of plastic debris. I ahve added a summary at the end of the code with refernce to a paper which adds more credibility to both the modelling and our thoughts and suggestions over the past meetings especially regarding the dataset and total evidence of sampling of all datasets we have encountered so far and wether there is bias in them.